// src/app/model/consumer-business.ts
export class ConsumerBusiness {
  firstName: string = '';
  lastName: string = '';
  dateOfBirth: Date = new Date();
  email: string = '';
  pan: string = '';
  businessOverview: string = '';
  businessType: string = '';
  annualTurnover: number = 0;
  totalEmployees: number = 0;
  capitalInvested: number = 0;
  businessAge: number = 0;
  agentUsername: string = '';
  constructor() {}
}